ComponentePorTerminar

+++ database/migrations/2025_04_09_152344_create_inventories_table.php
@@ -9,7 +9,7 @@ class CreateInventoriesTable extends Migration
     {
         Schema::create('inventories', function (Blueprint $table) {
             $table->id();
-            $table->Integer('user_id'); 
+            $table->unsignedBigInteger('user_id'); 
             $table->unsignedBigInteger('product_id'); 
             $table->integer('quantity'); 
             $table->timestamps();
